Move -help methods & license info into separate file
