+++ divi/src/LicenseAndInfo.h
@@ -1,0 +1,35 @@
+#ifndef LICENSE_AND_INFO_H
+#define LICENSE_AND_INFO_H
+#include <string>
+class CChainParams;
+
+/** The help message mode determines what help message to show */
+enum HelpMessageMode {
+    HMM_BITCOIND,
+    HMM_BITCOIN_QT
+};
+
+/** Help for options shared between UI and daemon (for -help) */
+std::string HelpMessage(HelpMessageMode mode, const CChainParams& defaultParameters);
+/** Returns licensing information (for -version) */
+std::string LicenseInfo();
+
+/**
+ * Format a string to be used as group of options in help messages
+ *
+ * @param message Group name (e.g. "RPC server options:")
+ * @return the formatted string
+ */
+std::string HelpMessageGroup(const std::string& message);
+
+/**
+ * Format a string to be used as option description in help messages
+ *
+ * @param option Option message (e.g. "-rpcuser=<user>")
+ * @param message Option description (e.g. "Username for JSON-RPC connections")
+ * @return the formatted string
+ */
+std::string HelpMessageOpt(const std::string& option, const std::string& message);
+
+std::string HelpMessageCli();
+#endif// LICENSE_AND_INFO_H
