+++ divi/src/init.h
@@ -20,16 +20,4 @@ bool ShutdownRequested();
 void Shutdown();
 void PrepareShutdown();
 bool InitializeDivi(boost::thread_group& threadGroup);
-
-/** The help message mode determines what help message to show */
-enum HelpMessageMode {
-    HMM_BITCOIND,
-    HMM_BITCOIN_QT
-};
-
-/** Help for options shared between UI and daemon (for -help) */
-std::string HelpMessage(HelpMessageMode mode);
-/** Returns licensing information (for -version) */
-std::string LicenseInfo();
-
 #endif // BITCOIN_INIT_H
